Added tests and second variant of the model.
I added variant of the model where ownership is fixed from the start and also a "decay" of votes for non-owners of a resource, named other_loss.
I also used the constants as parameters for the game, so they can be easily changed.
Constants PLAYERS and RESOURCES remain, I do not know if we want to change them in tests?
I further added some trivial test for the utility function and for the game, but there should be more for the game, I think.
I added tests for checking the pure strategies and they show that we do not always have pure strategies.
File data_large.csv gives data not for the regions, but for the towns. I wanted to check if we are able to apply the model to this number if we had only pure strategies, but it seemed, not surprisingly, to be super slow.